cloud/logsink/requirements.txt
flask==3.1.0
google-cloud-logging==3.13.0
gunicorn==23.0.0
